密码加密与微服务鉴权JWT模块结束(第9次)

+++ tensquare_parent/tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -9,7 +9,9 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
+import javax.servlet.http.HttpServletRequest;
 
+import io.jsonwebtoken.Claims;
 import net.bytebuddy.utility.RandomString;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -19,12 +21,15 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.http.HttpRequest;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import util.IdWorker;
 
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
+import util.JwtUtil;
 
 /**
  * 服务层
@@ -46,6 +51,15 @@ public class UserService {
 
 	@Autowired
 	private RabbitTemplate rabbitTemplate;
+
+	@Autowired
+	private BCryptPasswordEncoder encoder;
+
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@Autowired
+	private HttpServletRequest request;
 	/**
 	 * 查询全部列表
 	 * @return
@@ -94,6 +108,14 @@ public class UserService {
 	 */
 	public void add(User user) {
 		user.setId( idWorker.nextId()+"" );
+		//密码加密  spring的一个算法
+		user.setPassword(encoder.encode(user.getPassword()));
+		user.setFollowcount(0);
+		user.setFanscount(0);
+		user.setOnline(0L);
+		user.setRegdate(new Date());
+		user.setUpdatedate(new Date());
+		user.setLastdate(new Date());
 		userDao.save(user);
 	}
 
@@ -106,10 +128,35 @@ public class UserService {
 	}
 
 	/**
-	 * 删除
+	 * 删除  要有admin角色才可以 删除用户
 	 * @param id
 	 */
 	public void deleteById(String id) {
+		//以下验证 已经由拦截器处理了
+		/*//获取请求头 请求头里面放令牌
+		String header = request.getHeader("Authorization");
+		if (header == null || "".equals(header)){
+			throw  new RuntimeException("权限不足!");
+		}
+		//前后端约定：前端请求微服务时需要添加头信息Authorization ,内容为Bearer+空格 +token
+		//头部信息 key:Authorization  value:Bearer+空格
+		if (!header.startsWith("Bearer ")){
+			throw  new RuntimeException("权限不足!");
+		}
+		String token = header.substring(7);//截取得到token令牌
+		try {
+			Claims claims = jwtUtil.parseJWT(token);//解析token
+			String roles = (String) claims.get("roles");
+			if (roles == null || !roles.equals("admin")){
+				throw  new RuntimeException("权限不足!");
+			}
+		}catch (Exception e){
+			throw  new RuntimeException("权限不足!");
+		}*/
+		String token = (String) request.getAttribute("claims_admin");
+		if ("".equals(token)){
+			throw new RuntimeException("权限不足!");
+		}
 		userDao.deleteById(id);
 	}
 
@@ -179,8 +226,16 @@ public class UserService {
 		map.put("mobile",mobile);//手机号
 		map.put("code",checkCode);//验证码
 		//发送给用户一份
-		rabbitTemplate.convertAndSend("sms",map);
+		//rabbitTemplate.convertAndSend("sms",map);
 		System.out.println(checkCode);
     }
 
+    //用户登录
+	public User login(String mobile,String password) {
+		User user = userDao.findByMobile(mobile);//根据手机号查询出 用户对象
+		if (user!=null && encoder.matches(password,user.getPassword())){
+			return user;
+		}
+		return null;
+	}
 }
